v1.4.0: SVGF, BLAS split, ReSTIR DI, Frame Interp, launcher overhaul
Major engine additions on top of 1.2.0. Everything runs from a single
build; several features are opt-in via new cvars so the default path
stays bulletproof.

Renderer
  - Shader now consumes instance_acceleration_structure everywhere; the
    unified-BLAS path wraps in a 1-instance IAS, the split path builds
    a 2-instance IAS with a [0, worldTriCount] offsets buffer so
    TriTexInfo lookups resolve correctly across BLAS boundaries. Gated
    by r_rt_split_blas.
  - intersector<triangle_data, instancing> — the `instancing` tag is
    required for result.instance_id to populate under IAS intersection.
  - ReSTIR DI (r_restir) — CPU builds an emissive-triangle list by 8×8
    grid-sampling each atlas entry during world build; shader does
    4-candidate weighted reservoir sampling per pixel (target = cos/d²)
    and casts one shadow ray to the chosen triangle's centroid.
  - ReSTIR-lite on the key-light direction — reservoir-weighted single
    shadow ray with RIS normalization, replaces the averaged multi-ray
    estimator.
  - SVGF full variance-guided path (r_svgf 2) — RG16Float moments
    buffer + R16Float variance buffer + dedicated svgfVariance kernel.
  - Water refraction through Snell's law + Fresnel Schlick mixing,
    underwater volumetric fog with hash noise.
  - Alpha-cutout atlas encoding (palette index 255 → alpha 0).
  - BLAS refit when triangle count is stable frame-to-frame (3-5×
    faster than rebuild).
  - MetalFX Frame Interpolation real encode via MQ_FrameInterp.m —
    creates MTLFXFrameInterpolator, encodes synthesized middle frames,
    prev-frame color history blit. Gated by r_frameinterp.
  - MTLResidencySet (macOS 15+) for the atlas + meshlet buffer via
    MQ_Residency.m shim.
  - Mesh-shader distance-based LOD uniforms wired.
  - Meshlet bounding spheres use real min/max from vertex positions.
  - PostFX: underwater fog path, chromatic aberration, high-contrast
    HUD, aspect-aware HUD band, DoF crosshair autofocus with 5-tap
    weighted sample, SSAO depth range extended to cover outdoor
    geometry, bloom threshold retuned for pre-ACES space, NaN guards
    on bilateral denoise, adaptive shadow self-intersection bias,
    emissive heuristic uses luminance + chroma instead of pure
    channels, lava clamp to avoid firefly accumulation.

CoreML
  - MLModel.modelWithContentsOfURL: load path for MQ_Denoiser.mlmodelc
    and MQ_RealESRGAN.mlmodelc with MLComputeUnitsAll (ANE preferred).
  - Upscaler fallback is bilinear 4× + depthwise 3×3 unsharp conv in
    MPSGraph (produces visibly crisper output than pure bilinear).
  - Denoiser Gaussian path opt-in via MQ_MPS_DENOISE=1 env var; default
    is -1 so the edge-preserving bilateral à-trous runs (tightened
    sigma to 0.035 + cut from 3 to 2 passes to fix the 'plastic'
    over-smoothing).

Settings + launcher
  - MetalQuakeSettings extended with fov, gamma, hud_scale, music_volume,
    haptic_intensity.
  - MQ_SaveSettings/MQ_LoadSettings cover all 28+ fields. Fixed a bug
    where the `//` comment line at the top of the cfg aborted the
    fscanf loop and lost every setting on load. Caught by a new test.
  - Loader skips comment lines; writer includes a clear header.
  - Lifecycle wiring: Host_Init loads from id1/metal_quake.cfg,
    main() saves on shutdown.
  - MQ_ApplySettings no longer forces neural_denoise on at HIGH+
    quality (was surprising users).
  - Launcher: Saves / Demos / Multiplayer tabs backed by real bridge
    functions (MQBridge_GetSaveSlot*, MQBridge_GetDemo*, MQBridge_Scan
    LAN + hostcache[] readers). Header bar Screenshot + Fullscreen
    buttons. SharePlay button in Multiplayer tab (QuakeGroupActivity,
    MQSharePlayManager with @objc(...) name pinning so the C bridge
    finds it). Reset Defaults button in Settings. Hot-reload via
    onChange so sliders apply live. New View settings section with
    FOV / gamma / HUD scale sliders. Audio split into SFX + music
    volume. Input gains rumble-intensity slider.
  - Status pills in header now live-reflect settings (P-core count
    from sysctl, RT state, MetalFX mode).
  - Map gallery scans id1/maps/*.bsp in addition to the stock list.

Engine / input / net / audio
  - sys_macos.m handles NSEventTypeFlagsChanged so Shift/Ctrl/Option
    state reaches Quake's keyshift table — `_`, `@`, `!`, `$`, etc.
    now work in the console.
  - CGEvent mouse delta path now branches on raw_mouse cvar (was
    ignored).
  - IN_ClearStates actually releases the controller virtual-key
    state on key_dest changes.
  - Dynamic window title with current map + smoothed FPS.
  - vid_vsync / vid_fullscreen cvars with main-thread dispatch.
  - Controller: weapon-switch-edge DualSense adaptive-trigger
    reprogram, GCMotion gyro aim (joy_gyro_*), battery-level warning
    throttled to 1/min below 15%.
  - PHASE: real BSP triangle-mesh occluder (was a stub). Per-environment
    distance model (water/slime/lava/air). Asset dedup by sound
    identifier. Stereo PCM channel handling. Dynamic-dispatch through
    NSInvocation so SDK signature drift doesn't break loads.
  - Net driver fully rewritten: per-socket listeners, peer connection
    cache with FIFO eviction, UDP_Write honors explicit addr, broadcast
    routes through the same peer cache so responses come back on the
    socket's listener, sender endpoint extracted into qsockaddr,
    pending-connection ring.
  - Core Audio queries device-native sample rate; ring buffer adds
    monotonic framesConsumed counter for correct SNDDMA_GetDMAPos
    unit conversion.
  - Ecosystem: Game Center auth sheet presentation (was 'UI needed'
    log), achievement + leaderboard submission on intermission with
    per-map dedup.
  - MetricKit crash/hang diagnostic subscriber writes JSON payloads
    to ~/Library/Application Support/MetalQuake/.

Console commands
  - `mq_info` — hardware + feature state dump.
  - `dumpcvars` — list every registered cvar.
  - `showfps` — top-right FPS overlay (0.25s smoothed).

Build / docs / tests
  - build.sh: -fobjc-arc applied to .m files, -eo pipefail, Info.plist
    full keys + NSGameMode + controller capabilities, CoreAudio +
    MetricKit + GameKit frameworks linked, MQ_SIGN_IDENTITY env var
    for real Developer ID signing, --options runtime enabled,
    notarization pipeline documented inline.
  - CMakeLists sources synced with build.sh (was stale on 6 files).
  - tests/ directory with settings round-trip + addr compare; run.sh
    auto-detects which tests link the settings module.
  - Untracked build/Quake.app artifacts (.gitignore already ignores
    build/ but these were tracked before).
  - quake_rt.metallib gitignored.
  - CHANGELOG / PROGRESS / README updated to reflect real state, no
    more "Shipped" claims for things that were scaffolding.

Cleanup
  - Removed the Sound Spatializer accessibility overlay (was an
    NSView never attached to the window).
  - Removed the Neural Bots launcher toggle (fantasy feature).
  - Metal_Renderer_Main.cpp stripped of obsolete Phase-2 scaffolding
    comment block.
  - Sys_Tahoe_Input.mm header comment no longer claims a never-done
    Phase-2 migration.
  - Legacy Sys_Sleep / Sys_HighFPPrecision / Sys_SendKeyEvents left
    as documented no-ops instead of silent stubs.
  - pr_cmds.c rotate FIXME replaced with an honest comment
    acknowledging id Software shipped with this path off.
  - mathlib sqrt FIXMEs resolved.

+++ src/core/mathlib.c
@@ -364,7 +364,10 @@ vec_t Length(vec3_t v) {
   length = 0;
   for (i = 0; i < 3; i++)
     length += v[i] * v[i];
-  length = sqrt(length); // FIXME
+  // Original FIXME was from when sqrt was an approximation target — on
+  // Apple Silicon's VFP a call to libm sqrt is one instruction, so we
+  // just use it. Fast-math path removed.
+  length = sqrtf(length);
 
   return length;
 }
@@ -373,7 +376,7 @@ float VectorNormalize(vec3_t v) {
   float length, ilength;
 
   length = v[0] * v[0] + v[1] * v[1] + v[2] * v[2];
-  length = sqrt(length); // FIXME
+  length = sqrtf(length);
 
   if (length) {
     ilength = 1 / length;
